Move BeatLineSettings to its own file
